LayoutTests/media/video-src-empty-expected.txt

Empty src attribute defined. Invalid URL and video should error.

EXPECTED (video.error.code == '4') OK
EXPECTED (video.networkState == '4') OK
END OF TEST
